+++ OUDS/Core/Components/Sources/Controls/TextArea/Internal/TextAreaContainer.swift
@@ -124,6 +124,8 @@ struct TextAreaContainer: View {
                 "core_common_disabled_a11y".localized()
             case let .error(message):
                 "core_common_onError_a11y".localized() + ": \(message)"
+            case let .richError(message):
+                "core_common_onError_a11y".localized() + ": \(String(message.characters))"
             case .loading:
                 "core_common_loading_a11y".localized()
             case .enabled: